Relocate data methods to score calculation data service
